refactor: move functions depend on spark
