+++ openaivec/util.py
@@ -4,12 +4,11 @@ import time
 from concurrent.futures.thread import ThreadPoolExecutor
 from dataclasses import dataclass
 from itertools import chain
-from typing import Callable, List, Optional, Type, TypeVar, Union, get_args, get_origin
+from typing import Callable, List, Optional, TypeVar
 
 import numpy as np
 import tiktoken
-from pydantic import BaseModel
-from pyspark.sql.types import ArrayType, BooleanType, FloatType, IntegerType, StringType, StructField, StructType
+
 
 T = TypeVar("T")
 U = TypeVar("U")
@@ -71,51 +70,6 @@ def map_unique_minibatch_parallel(b: List[T], batch_size: int, f: Callable[[List
     and the results are mapped back to match the order of the original list.
     """
     return map_unique(b, lambda x: map_minibatch_parallel(x, batch_size, f))
-
-
-def python_type_to_spark(python_type):
-    origin = get_origin(python_type)
-
-    # For list types (e.g., List[int])
-    if origin is list or origin is List:
-        # Retrieve the inner type and recursively convert it
-        inner_type = get_args(python_type)[0]
-        return ArrayType(python_type_to_spark(inner_type))
-
-    # For Optional types (Union[..., None])
-    elif origin is Union:
-        non_none_args = [arg for arg in get_args(python_type) if arg is not type(None)]
-        if len(non_none_args) == 1:
-            return python_type_to_spark(non_none_args[0])
-        else:
-            raise ValueError(f"Unsupported Union type with multiple non-None types: {python_type}")
-
-    # For nested Pydantic models (to be treated as Structs)
-    elif isinstance(python_type, type) and issubclass(python_type, BaseModel):
-        return pydantic_to_spark_schema(python_type)
-
-    # Basic type mapping
-    elif python_type is int:
-        return IntegerType()
-    elif python_type is float:
-        return FloatType()
-    elif python_type is str:
-        return StringType()
-    elif python_type is bool:
-        return BooleanType()
-    else:
-        raise ValueError(f"Unsupported type: {python_type}")
-
-
-def pydantic_to_spark_schema(model: Type[BaseModel]) -> StructType:
-    fields = []
-    for field_name, field in model.model_fields.items():
-        field_type = field.annotation
-        # Use outer_type_ to correctly handle types like Optional
-        spark_type = python_type_to_spark(field_type)
-        # Set nullable to True (adjust logic as needed)
-        fields.append(StructField(field_name, spark_type, nullable=True))
-    return StructType(fields)
 
 
 def get_exponential_with_cutoff(scale: float) -> float:
